refactor, pass tests with ruby TLC
