+++ spec/site/alarm_spec.rb
@@ -1,9 +1,9 @@
 RSpec.describe 'RSMP site alarm' do
-  it 'A0302 detector error (logic error)' do |example|
+  it 'A0302 detector error (logic error)', script: true do |example|
     TestSite.log_test_header example
     TestSite.connected do |task,supervisor,site|
       component = COMPONENT_CONFIG['detector_logic'].keys.first
-      system("~/activate_alarm.sh")
+      system(SCRIPT_PATHS['activate_alarm'])
       site.log "Waiting for alarm", level: :test
       start_time = Time.now
       message, response = nil,nil
@@ -14,18 +14,27 @@ RSpec.describe 'RSMP site alarm' do
 
       delay = Time.now - start_time
       site.log "alarm confirmed after #{delay.to_i}s", level: :test
-      system("~/deactivate_alarm.sh")
+      system(SCRIPT_PATHS['deactivate_alarm'])
 
       alarm_time = Time.parse(response[:message].attributes["aTs"])
       expect(alarm_time).to be_within(1.minute).of Time.now.utc
+      expect(response[:message].attributes['rvs']).to eq([{
+        "n":"detector","v":"1"},
+        {"n":"type","v":"loop"},
+        {"n":"errormode","v":"on"},
+        {"n":"manual","v":"True"},
+        {"n":"logicerror","v":"always_off"}
+      ])
+    ensure
+      system(SCRIPT_PATHS['deactivate_alarm'])
     end
   end
 
-  it 'Acknowledge alarm' do |example|
+  it 'Acknowledge alarm', script: true do |example|
     TestSite.log_test_header example
     TestSite.connected do |task,supervisor,site|
       component = COMPONENT_CONFIG['detector_logic'].keys.first
-      system("~/activate_alarm.sh")
+      system(SCRIPT_PATHS['activate_alarm'])
       site.log "Waiting for alarm", level: :test
       start_time = Time.now
       message, response = nil,nil
@@ -41,14 +50,14 @@ RSpec.describe 'RSMP site alarm' do
       site.log "alarm confirmed after #{delay.to_i}s", level: :test
       
       expect do
-        response = @site.wait_for_alarm_acknowledged_response message: message, component: @component, timeout: RSMP_CONFIG['command_timeout']
+        response = @site.wait_for_alarm_acknowledged_response message: message, component: @component, timeout: RSMP_CONFIG['alarm_timeout']
       end.to_not raise_error
       
       expect(response).not_to be_a(RSMP::MessageNotAck), "Message rejected: #{response.attributes['rea']}"
       expect(response).to be_a(RSMP::AlarmAcknowledgedResponse)
       expect(response.attributes['cId']).to eq(@component)
-      
-      system("~/deactivate_alarm.sh")
+    ensure 
+      system(SCRIPT_PATHS['deactivate_alarm'])
     end
   end
 end
